pyproject.toml
[project]
name = "langchain-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "gitpython>=3.1.45",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-openai>=0.3.35",
    "langchain-text-splitters>=0.3.11",
    "python-dotenv>=1.1.1",
]
